Refactored `removeBlocksWithCallback`
